feat(ordonnancement/postier): #29 priorité décimale, #44 colonne dispo semi, #47 EQC chargés, #54 volumes disponibles, #56 dates de l'historique
#29 — PRIORITÉ DÉCIMALE
L'agent trafic veut intercaler une tournée entre la 2 et la 3 sans renuméroter
les autres. La colonne `tours.priority` était un entier : 2,5 était arrondi
silencieusement. Elle a été élargie en numeric(8,2) par une migration manuelle
en production — la migration automatique du démarrage ajoute des colonnes mais
n'altère pas les types. Conversion contrôlée avant/après : 1 099 priorités
conservées, somme identique (21 227). Le champ de saisie accepte désormais les
décimales, et l'export « Ordre » écrit un nombre et non une chaîne.

#44 — COLONNE DISPO SEMI
L'heure de disponibilité de la semi n'était lisible qu'en dépliant la tournée.
Elle devient une colonne de la liste postier, masquable comme les autres.

#47 — CASE EQC CHARGÉS
Le champ était en lecture seule, alimenté par le seul import WMS : quand le
chargement réel ne correspondait pas, le postier n'avait aucun moyen de
corriger. Il devient saisissable, l'import continuant de le pré-remplir.
Le champ est un entier en base : la saisie est donc contrainte à un nombre
entier plutôt que d'accepter des décimales qui seraient tronquées en silence.

#54 — PIOCHER DANS LES VOLUMES DISPONIBLES
Ajouter un point de vente à une tournée validée imposait de taper les EQC à
l'aveugle. La recherche affiche maintenant, pour chaque point de vente, les EQC
encore non attribués de la journée, remonte ceux qui en ont en tête de liste et
pré-remplit la quantité. Le rattachement des volumes était déjà fait par le
serveur. Le second volet du ticket — modifier l'ordre des PDV d'une tournée
validée — a été livré avec le #37.

#56 — DATES DE L'HISTORIQUE
La colonne « Date » portait la date de RÉPARTITION sans le dire. Elle est
renommée, et la date du tour (livraison) devient une colonne filtrable à côté.

AU PASSAGE — TRONCATURE SILENCIEUSE DANS L'ÉCRAN POSTIER
Les volumes y étaient chargés sans aucun filtre de date : la requête héritait du
plafond serveur de 500 lignes, prises au hasard dans les 15 000 de la base. Les
températures et dates de répartition affichées par arrêt pouvaient donc être
muettes — et le calcul des EQC disponibles ajouté pour le #54 aurait été faux.
Seules les journées de répartition des tournées affichées sont désormais
chargées : une ou deux requêtes ciblées. Même famille que #27 et #83.

Tests : tests/test_priority_decimal.py (3) ; 182 passed, isolation 14 passed,
build et lint front sans régression.

+++ Chaos_Route/backend/app/models/tour.py
@@ -86,7 +86,12 @@ class Tour(Base, TenantMixin):
     bypass_support_rules: Mapped[bool] = mapped_column(Boolean, default=False)  # Desactive le controle support/base pour cette tournee
     # Priorité manuelle d'ordonnancement (1..n) saisie au moment de planifier ;
     # départage les tours à même heure de départ. NULL = non prioritaire (en dernier).
-    priority: Mapped[int | None] = mapped_column(Integer)
+    # Priorité d'ordonnancement, DÉCIMALE (#29) : l'agent trafic intercale une
+    # tournée entre la 2 et la 3 en saisissant 2,5, sans renuméroter les autres.
+    # La colonne était un entier ; elle a été élargie en numeric(8,2) par une
+    # migration manuelle (la migration de démarrage n'altère pas les types).
+    # / Decimal scheduling priority so a tour can be slotted between two others.
+    priority: Mapped[float | None] = mapped_column(Numeric(8, 2))
 
     # Champs opérationnels / Operational fields — datetime-local YYYY-MM-DDTHH:MM
     driver_name: Mapped[str | None] = mapped_column(String(100))
